+++ params/avalanche_params.go
@@ -5,7 +5,6 @@ package params
 
 import (
 	"github.com/ava-labs/avalanchego/utils/units"
-	"github.com/ava-labs/avalanchego/utils/wrappers"
 )
 
 // Minimum Gas Price
@@ -19,14 +18,5 @@ const (
 
 	ApricotPhase1GasLimit uint64 = 8_000_000
 	CortinaGasLimit       uint64 = 15_000_000
-
-	ApricotPhase3MinBaseFee               int64 = 75 * GWei
-	ApricotPhase3MaxBaseFee                     = 225 * GWei
-	ApricotPhase3InitialBaseFee           int64 = 225 * GWei
-	ApricotPhase3TargetGas                      = 10_000_000
-	ApricotPhase3BaseFeeChangeDenominator       = 12
-	EtnaMinBaseFee                        int64 = GWei
-
-	RollupWindow            = 10 // in seconds
-	DynamicFeeExtraDataSize = wrappers.LongLen * RollupWindow
+	EtnaMinBaseFee        int64  = GWei
 )
