+++ plugin/evm/api.go
@@ -15,7 +15,7 @@ import (
 	"github.com/ava-labs/avalanchego/utils/formatting"
 	"github.com/ava-labs/avalanchego/utils/json"
 	"github.com/ava-labs/avalanchego/utils/set"
-	"github.com/ava-labs/coreth/params"
+	"github.com/ava-labs/coreth/plugin/evm/ap3"
 	"github.com/ava-labs/coreth/plugin/evm/atomic"
 	"github.com/ava-labs/coreth/plugin/evm/client"
 	"github.com/ethereum/go-ethereum/common"
@@ -36,7 +36,7 @@ var (
 	errNoSourceChain = errors.New("no source chain provided")
 	errNilTxID       = errors.New("nil transaction ID")
 
-	initialBaseFee = big.NewInt(params.ApricotPhase3InitialBaseFee)
+	initialBaseFee = big.NewInt(ap3.InitialBaseFee)
 )
 
 // SnowmanAPI introduces snowman specific functionality to the evm
