+++ plugin/evm/header/extra.go
@@ -7,10 +7,31 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/ava-labs/coreth/core/types"
 	"github.com/ava-labs/coreth/params"
 )
 
 var errInvalidExtraLength = errors.New("invalid header.Extra length")
+
+// ExtraPrefix takes the previous header and the timestamp of its child block
+// and calculates the expected extra prefix for the child block.
+func ExtraPrefix(
+	config *params.ChainConfig,
+	parent *types.Header,
+	timestamp uint64,
+) ([]byte, error) {
+	switch {
+	case config.IsApricotPhase3(timestamp):
+		window, err := feeWindow(config, parent, timestamp)
+		if err != nil {
+			return nil, fmt.Errorf("failed to calculate fee window: %w", err)
+		}
+		return feeWindowBytes(window), nil
+	default:
+		// Prior to AP3 there was no expected extra prefix.
+		return nil, nil
+	}
+}
 
 // VerifyExtra verifies that the header's Extra field is correctly formatted for
 // [rules].
@@ -18,20 +39,20 @@ func VerifyExtra(rules params.AvalancheRules, extra []byte) error {
 	extraLen := len(extra)
 	switch {
 	case rules.IsDurango:
-		if extraLen < params.DynamicFeeExtraDataSize {
+		if extraLen < FeeWindowSize {
 			return fmt.Errorf(
 				"%w: expected >= %d but got %d",
 				errInvalidExtraLength,
-				params.DynamicFeeExtraDataSize,
+				FeeWindowSize,
 				extraLen,
 			)
 		}
 	case rules.IsApricotPhase3:
-		if extraLen != params.DynamicFeeExtraDataSize {
+		if extraLen != FeeWindowSize {
 			return fmt.Errorf(
 				"%w: expected %d but got %d",
 				errInvalidExtraLength,
-				params.DynamicFeeExtraDataSize,
+				FeeWindowSize,
 				extraLen,
 			)
 		}
